[iOS] Fix: CoreData 모델 파일이 안열리던 문제 해결, 파일 path 수정
