+++ MiniVibe/MiniVibe/CoreData/Like+CoreDataClass.swift
